requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
pydantic==2.9.2
pymilvus==2.6.2
jinja2==3.1.4
anyio==4.4.0
sentence-transformers==3.0.1
marshmallow==3.20.1
python-dotenv==1.1.1
pydantic-settings==2.11.0
python-docx==1.1.2
PyPDF2==3.0.1
pandas==2.2.2
openpyxl==3.1.5
python-multipart==0.0.20
jose==1.0.0
ldap3==2.9.1
python-jose[cryptography]
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.28.0
redis>=5.0.0